server_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-google-genai==2.0.7
langchain-core==0.3.18
pydantic==2.9.2
python-multipart==0.0.12
azure-cognitiveservices-speech==1.40.0
pydub==0.25.1
google-cloud-texttospeech==2.21.2
openai==1.54.3
requests==2.32.3
feedparser==6.0.11

